final review improvements: - added cache - changed money to use long - fixed minor issues and code smells - updated junits

+++ src/main/java/com/haiilo/kata/controller/SupermarketCheckoutController.java
@@ -20,7 +20,7 @@ public class SupermarketCheckoutController {
 
     @PostMapping("checkout")
     public ResponseEntity<CheckoutResponse> checkout(@Valid @RequestBody CheckoutRequest checkoutRequest) {
-        int totalPriceInMinor = priceService.calculateTotalPrice(checkoutRequest);
+        long totalPriceInMinor = priceService.calculateTotalPrice(checkoutRequest);
         return ResponseEntity.ok(new CheckoutResponse(totalPriceInMinor));
     }
 }
